# promql: Hash two labelsets together