--- a/promql/engine.go
+++ b/promql/engine.go
@@ -830,9 +830,8 @@ type EvalNodeHelper struct {
 	regex *regexp.Regexp
 
 	// For binary vector matching.
-	rightSigs    map[uint64]Sample
-	matchedSigs  map[uint64]map[uint64]struct{}
-	resultMetric map[uint64]labels.Labels
+	rightSigs   map[uint64]Sample
+	matchedSigs map[uint64]map[uint64]struct{}
 }
 
 // dropMetricName is a cached version of dropMetricName.
@@ -1692,19 +1691,6 @@ func signatureFunc(on bool, names ...string) func(labels.Labels) uint64 {
 // resultMetric returns the metric for the given sample(s) based on the Vector
 // binary operation and the matching options.
 func resultMetric(lhs, rhs labels.Labels, op ItemType, matching *VectorMatching, enh *EvalNodeHelper) labels.Labels {
-	if enh.resultMetric == nil {
-		enh.resultMetric = make(map[uint64]labels.Labels, len(enh.out))
-	}
-	// op and matching are always the same for a given node, so
-	// there's no need to include them in the hash key.
-	// If the lhs and rhs are the same then the xor would be 0,
-	// so add in one side to protect against that.
-	lh := lhs.Hash()
-	h := (lh ^ rhs.Hash()) + lh
-	if ret, ok := enh.resultMetric[h]; ok {
-		return ret
-	}
-
 	lb := labels.NewBuilder(lhs)
 
 	if shouldDropMetricName(op) {
@@ -1736,7 +1722,6 @@ func resultMetric(lhs, rhs labels.Labels, op ItemType, matching *VectorMatching,
 	}
 
 	ret := lb.Labels()
-	enh.resultMetric[h] = ret
 	return ret
 }