@roidelapluie 2020-01-04
Benchmark (go test  -benchtime=60s -run NONE -test.benchmem=true -bench=.)

```
benchmark                     old ns/op     new ns/op     delta
BenchmarkSignature-8          584           559           -4.28%
BenchmarkSignatureReuse-8     128           102           -20.31%

benchmark                     old allocs     new allocs     delta
BenchmarkSignature-8          7              7              +0.00%
BenchmarkSignatureReuse-8     1              1              +0.00%

benchmark                     old bytes     new bytes     delta
BenchmarkSignature-8          672           672           +0.00%
BenchmarkSignatureReuse-8     32            32            +0.00%
```