@brian-brazil 2020-01-05
I think we should handle collisions correctly, they're going to happen no matter how good the hash function is.