@juliusv 2020-01-05
Agreed with @brian-brazil but OTOH if this makes the existing collisions way less likely it's at least a better stop-gap than nothing. But yeah, the end goal should be to have collision handling.