@roidelapluie 2020-01-05
@brian-brazil is that what you mean?

```
benchmark                                                                                                                                                                                                                                                                                                                old ns/op     new ns/op     delta
BenchmarkSignature/{}{}-8                                                                                                                                                                                                                                                                                                468           556           +18.80%
BenchmarkSignature/{job="prometheus",___name__="up"}{}-8                                                                                                                                                                                                                                                                 584           653           +11.82%
BenchmarkSignature/{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}{}-8                                                                                                                                                                                                                                    594           671           +12.96%
BenchmarkSignature/{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}{}-8                                                                                                                                                     791           838           +5.94%
BenchmarkSignature/{}{job="prometheus",___name__="up"}-8                                                                                                                                                                                                                                                                 533           574           +7.69%
BenchmarkSignature/{job="prometheus",___name__="up"}{job="prometheus",___name__="up"}-8                                                                                                                                                                                                                                  629           668           +6.20%
BenchmarkSignature/{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}{job="prometheus",___name__="up"}-8                                                                                                                                                                                                     623           705           +13.16%
BenchmarkSignature/{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}{job="prometheus",___name__="up"}-8                                                                                                                      772           861           +11.53%
BenchmarkSignature/{}{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}-8                                                                                                                                                                                                                                    504           581           +15.28%
BenchmarkSignature/{job="prometheus",___name__="up"}{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}-8                                                                                                                                                                                                     628           674           +7.32%
BenchmarkSignature/{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}-8                                                                                                                                                                        637           681           +6.91%
BenchmarkSignature/{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}-8                                                                                         804           869           +8.08%
BenchmarkSignature/{}{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}-8                                                                                                                                                     582           636           +9.28%
BenchmarkSignature/{job="prometheus",___name__="up"}{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}-8                                                                                                                      663           736           +11.01%
BenchmarkSignature/{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}-8                                                                                         681           759           +11.45%
BenchmarkSignature/{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}-8          854           910           +6.56%
BenchmarkSignatureReuse/{}{}-8                                                                                                                                                                                                                                                                                           58.6          49.2          -16.04%
BenchmarkSignatureReuse/{job="prometheus",___name__="up"}{}-8                                                                                                                                                                                                                                                            85.5          81.6          -4.56%
BenchmarkSignatureReuse/{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}{}-8                                                                                                                                                                                                                               91.1          89.7          -1.54%
BenchmarkSignatureReuse/{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}{}-8                                                                                                                                                120           127           +5.83%
BenchmarkSignatureReuse/{}{job="prometheus",___name__="up"}-8                                                                                                                                                                                                                                                            88.0          80.7          -8.30%
BenchmarkSignatureReuse/{job="prometheus",___name__="up"}{job="prometheus",___name__="up"}-8                                                                                                                                                                                                                             115           112           -2.61%
BenchmarkSignatureReuse/{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}{job="prometheus",___name__="up"}-8                                                                                                                                                                                                121           115           -4.96%
BenchmarkSignatureReuse/{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}{job="prometheus",___name__="up"}-8                                                                                                                 152           160           +5.26%
BenchmarkSignatureReuse/{job="prometheus",___name__="up"}{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}-8                                                                                                                                                                                                128           126           -1.56%
BenchmarkSignatureReuse/{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}-8                                                                                                                                                                   133           128           -3.76%
BenchmarkSignatureReuse/{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}-8                                                                                    162           168           +3.70%
BenchmarkSignatureReuse/{}{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}-8                                                                                                                                                130           138           +6.15%
BenchmarkSignatureReuse/{job="prometheus",___name__="up"}{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}-8                                                                                                                 157           172           +9.55%
BenchmarkSignatureReuse/{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}-8                                                                                    162           169           +4.32%
BenchmarkSignatureReuse/{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}-8     194           212           +9.28%

benchmark                                                                                                                                                                                                                                                                                                           old bytes     new bytes     delta
BenchmarkSignature/{}{}-8                                                                                                                                                                                                                                                                                           608           1024          +68.42%
BenchmarkSignature/{job="prometheus",___name__="up"}{}-8                                                                                                                                                                                                                                                            672           1088          +61.90%
BenchmarkSignature/{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}{}-8                                                                                                                                                                                                                               672           1088          +61.90%
BenchmarkSignature/{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}{}-8                                                                                                                                                768           1184          +54.17%
BenchmarkSignature/{}{job="prometheus",___name__="up"}-8                                                                                                                                                                                                                                                            608           1024          +68.42%
BenchmarkSignature/{job="prometheus",___name__="up"}{job="prometheus",___name__="up"}-8                                                                                                                                                                                                                             672           1088          +61.90%
BenchmarkSignature/{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}{job="prometheus",___name__="up"}-8                                                                                                                                                                                                672           1088          +61.90%
BenchmarkSignature/{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}{job="prometheus",___name__="up"}-8                                                                                                                 768           1184          +54.17%
BenchmarkSignature/{}{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}-8                                                                                                                                                                                                                               608           1024          +68.42%
BenchmarkSignature/{job="prometheus",___name__="up"}{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}-8                                                                                                                                                                                                672           1088          +61.90%
BenchmarkSignature/{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}-8                                                                                                                                                                   672           1088          +61.90%
BenchmarkSignature/{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}-8                                                                                    768           1184          +54.17%
BenchmarkSignature/{}{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}-8                                                                                                                                                608           1024          +68.42%
BenchmarkSignature/{job="prometheus",___name__="up"}{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}-8                                                                                                                 672           1088          +61.90%
BenchmarkSignature/{job="prometheus",___name__="go_memstats_buck_hash_sys_bytes"}{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}-8                                                                                    672           1088          +61.90%
BenchmarkSignature/{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}{instance="localhost:9090",_job="prometheus",___name__="prometheus_http_request_duration_seconds_bucket",_handler="/api/v1/*path",_le="+Inf"}-8     768           1184          +54.17%
```